tests/requirements_fast.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
faiss-cpu==1.7.4
sentence-transformers==2.2.2
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
streamlit==1.28.1
pypdf==3.17.1
tiktoken==0.5.2 